Move test files out of __tests__
