--- packages/module/patternfly-docs/content/extensions/react-console/examples/Basic.tsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import { ExtendedButton } from '@patternfly/react-console';
-
-export const BasicExample: React.FunctionComponent = () => <ExtendedButton>My custom extension button</ExtendedButton>;
