+++ packages/module/patternfly-docs/content/extensions/react-console/examples/SerialConsoleCustom.jsx
