+++ packages/module/src/components/VncConsole/VncActions.test.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import { VncActions } from '../VncActions';
+import { VncActions } from './VncActions';
 
 describe('VncActions', () => {
   test('placeholder render test', () => {
